renamed number-input -> NumberInput and global-event -> GlobalEvent
