+++ packages/wix-ui-core/src/components/global-event/GlobalEvent.tsx
