+++ packages/wix-ui-core/src/components/global-event/index.ts
@@ -1,1 +1,1 @@
-export * from './global-event';
+export * from './GlobalEvent';
