+++ packages/wix-ui-core/src/components/number-input/NumberInput.st.css
@@ -1,5 +1,5 @@
 :import {
-    -st-from: "./stepper.st.css";
+    -st-from: "./Stepper.st.css";
     -st-default: Stepper;
 }
