+++ packages/wix-ui-core/src/components/number-input/NumberInput.story.tsx
@@ -1,14 +1,18 @@
-import {NumberInput} from '../src/components/number-input'
+import {NumberInput} from './NumberInput'
 
 export default {
   category: 'Components',
   storyName: 'NumberInput',
 
   component: NumberInput,
-  componentPath: '../src/components/number-input/number-input.tsx',
+  componentPath: './NumberInput.tsx',
 
   componentProps: {
     value: 20,
+    placeholder: 'Enter number',
+    min: 4,
+    max: 34,
+    step: 2,
     'data-hook': 'storybook-button'
   },
   exampleProps: {
