+++ packages/wix-ui-core/src/components/number-input/Stepper.st.css
