+++ packages/wix-ui-core/src/components/number-input/index.ts
@@ -1,1 +1,1 @@
-export * from './number-input';
+export * from './NumberInput';
