--- packages/wix-ui-core/src/components/number-input/number-input.story.tsx
@@ -1,15 +1,0 @@
-import {NumberInput} from './';
-
-export default {
-  category: 'Components',
-  storyName: 'NumberInput',
-  component: NumberInput,
-  componentPath: './number-input.tsx',
-  componentProps: {
-    value: 20,
-    'data-hook': 'storybook-button'
-  },
-  exampleProps: {
-    onClick: () => 'Clicked!',
-  }
-}
